requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.7.0
scikit-learn>=0.24.2
tensorboard>=2.12.0

# Image Processing
Pillow>=9.5.0
opencv-python>=4.5.0
albumentations>=1.3.0

# OCR and Text Processing
pytesseract>=0.3.8
easyocr>=1.4.0
transformers>=4.15.0

# Web Framework and API
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0

# Data Management
pandas>=1.3.3
h5py>=3.6.0
tqdm>=4.65.0

# Visualization and Reporting
matplotlib>=3.5.0
seaborn>=0.11.0
reportlab>=3.6.1
grad-cam>=1.4.0

# Testing and Development
pytest>=6.2.0
black>=21.5b2
flake8>=3.9.0
isort>=5.9.0

# Utilities
python-dotenv>=0.19.0
requests>=2.26.0
scikit-image>=0.18.3
gdown>=4.6.0 